+++ packages/storybook/src/nimble/patterns/tabs/doc-strings.ts
@@ -1,0 +1,5 @@
+import { buttonTag } from '@ni/nimble-components/dist/esm/button';
+import { tabsToolbarTag } from '@ni/nimble-components/dist/esm/tabs-toolbar';
+
+export const defaultSlotDescription = `Slot to display \`${buttonTag}\` elements on the left side of the \`${tabsToolbarTag}\`. If this slot is empty, the separator will be hidden.`;
+export const endSlotDescription = `Slot to display \`${buttonTag}\` elements on the right side of the \`${tabsToolbarTag}\`.`;