+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-components
 
-<!-- This log was last generated on Tue, 10 Jun 2025 20:39:08 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 26 Jun 2025 19:20:55 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 33.11.1
+
+Thu, 26 Jun 2025 19:20:55 GMT
+
+### Patches
+
+- Fix issue with default slotted content in tabs toolbar that sets its margin-left to 'auto'. ([ni/nimble@0bbb3df](https://github.com/ni/nimble/commit/0bbb3dfc4dcdbcdfd80e31ea877c4386d93156c6))
+
+## 33.11.0
+
+Tue, 24 Jun 2025 16:53:02 GMT
+
+### Minor changes
+
+- Provide start/end slots for tabs toolbar. ([ni/nimble@198d94f](https://github.com/ni/nimble/commit/198d94f59fad1d05605048ade29ede19cb8b8c32))
+
+## 33.10.0
+
+Fri, 20 Jun 2025 06:07:44 GMT
+
+### Minor changes
+
+- Add theme aware token for calendar event outer border ([ni/nimble@3946d5c](https://github.com/ni/nimble/commit/3946d5cbe190276f9373a4d3b72ee47ee3681d75))
 
 ## 33.9.0